requirements.txt
aiogram==2.25.1
fastapi
uvicorn
python-dotenv==1.0.0
pillow==10.2.0
numpy==1.26.3
scikit-image
scikit-learn==1.3.0
gdown
